rename network-core to core
