+++ Cargo.toml
@@ -13,10 +13,10 @@ default = []
 opt-level = 3
 
 [workspace]
-members = ["network-core"]
+members = ["core"]
 
 [dependencies]
-network-core = { path = "./network-core", features = ["serde", "serde_json"] }
+core = { path = "./core", package = "fishfight-core", features = ["serde", "serde_json"] }
 # ultimate = { path = "../FishFight-ultimate", package = "fishfight-ultimate", optional = true }
 ff-particles = { version = "0.1", features = ["serde"] }
 macroquad = { version = "0.3.10" }
