+++ core/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "fishfight-core"
+version = "0.3.0"
+authors = ["Fish Fight Contributors"]
+license = "MIT OR Apache-2.0"
+edition = "2021"
+
+[features]
+default = []
+
+[dependencies]
+macroquad = { version = "0.3.10" }
+hecs = "0.7.1"
+serde = { version = "1.0", package = "serde", features = ["derive"], optional = true }
+serde_json = { version = "1.0", optional = true }
+async-trait = "0.1.52"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+fishsticks = { version = "0.2.0", features = ["bundled-sdl2"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+fishsticks = { version = "0.2.0", default-features = false, features = ["gilrs"] }
