+++ core/src/math.rs
