+++ core/src/text.rs
