--- network-core/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "network-core"
-version = "0.3.0"
-authors = ["Fish Fight Contributors"]
-license = "MIT OR Apache-2.0"
-edition = "2021"
-
-[features]
-default = []
-
-[dependencies]
-async-trait = "0.1.52"
-serde = { version = "1.0", package = "serde", features = ["derive"], optional = true }
-serde_json = { version = "1.0", optional = true }
