+++ src/effects/active/triggered.rs
@@ -5,11 +5,12 @@ use hecs::{Entity, World};
 
 use serde::{Deserialize, Serialize};
 
+use core::math::deg_to_rad;
+use core::Result;
+
 use crate::effects::active::spawn_active_effect;
-use crate::math::deg_to_rad;
 use crate::particles::{ParticleEmitter, ParticleEmitterMetadata};
 use crate::player::{Player, PlayerState};
-use crate::Result;
 use crate::{json, Drawable, PhysicsBodyParams};
 use crate::{ActiveEffectMetadata, AnimatedSpriteMetadata, CollisionWorld, PhysicsBody, Transform};
