+++ src/network.rs
@@ -2,16 +2,15 @@
 
 use hecs::World;
 
-pub use network_core::{Api, Id, Lobby, Player, DEFAULT_PORT};
-
-use crate::Result;
+use core::network::Api;
+use core::Result;
 
 pub async fn init_api(token: &str) -> Result<()> {
     #[cfg(feature = "ultimate")]
     Api::init::<ultimate::UltimateApiBackend>(token).await?;
 
     #[cfg(not(feature = "ultimate"))]
-    Api::init::<network_core::MockApiBackend>(token).await?;
+    Api::init::<core::network::MockApiBackend>(token).await?;
 
     Ok(())
 }
